refactor(documents): simplify processor architecture, add page selector UI
Backend:
- Remove DocumentProcessorManager, use direct processor instantiation
- Add AI_PROCESSOR env var (set to "gemini" to enable AI transform)
- API now uses ai_transform: bool instead of processor_slug
- Processors have hardcoded slugs and sensible defaults
- Remove document_processors.*.json config files
- Fix process_with_billing to accept metadata directly (no redundant cache lookup)

Frontend:
- Rename OCR toggle to "AI Transform"
- Add page selector (text input + visual bar) for multi-page PDFs
- Page selector works with both free and AI processing

+++ yapit/gateway/domain_models.py
@@ -70,7 +70,7 @@ class DocumentMetadata(PydanticModel):
     title: str | None = Field(default=None, max_length=500)
     url: str | None = Field(default=None, max_length=2000)
     file_name: str | None = Field(default=None, max_length=255)
-    file_size: float | None = None  # Content size in bytes
+    file_size: int | None = None  # Content size in bytes
 
 
 class Document(SQLModel, table=True):
